Try to make linter happy
